// generated/src/app/api-models/inst-claim-detail-edi-w.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class InstClaimDetailEdiW {
  seqPrediId: number;
  seqClaimId: number;
  claimNumber: string;
  originalClaimNumber: string;
  lineNumber: number;
  subLineCode: string;
  detailSvcDate: Date;
  detailThruDate: Date;
  diagnosisPointer: number;
  placeServPointer: number;
  seqProvId: number;
  seqMembId: number;
  billedProcedure: string;
  billedModifier: string;
  billedAllowedAmt: number;
  procedureCode: string;
  procedureModifier: string;
  quantity: number;
  billedAmt: number;
  allowedFactor: number;
  allowedAmt: number;
  notCoveredAmt: number;
  copayment1Amt: number;
  copayment2Amt: number;
  deductibleAmt: number;
  otherCarrierAmt: number;
  withholdAmt: number;
  netAmt: number;
  allowedReason: string;
  notCoveredReason: string;
  copay1Reason: string;
  copay2Reason: string;
  deductibleReason: string;
  adjustmentReason: string;
  otherCarrierRsn: string;
  holdReason1: string;
  holdReason2: string;
  holdReason3: string;
  claimStatus: string;
  processingStatus: string;
  adjudicationMethod: string;
  medDefCode: string;
  postDate: Date;
  checkDate: Date;
  companyCode: string;
  glRefCode: string;
  printFlag: string;
  eobId: string;
  raId: string;
  seqApTrans: number;
  seqAuthDetail: number;
  hiddenUserDef1: string;
  hiddenUserDef2: number;
  countAsDays: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  carveOut: string;
  systemGenerated: string;
  capOutlierInclExcl: string;
  ediStatus: string;
  alternateProcCode: string;
  accommodationRate: number;
  nationalUnlabeled1: string;
  withholdSurplus: number;
  recInd: string;
}